Move maintenance:mimetypesjs to sublocation, cleanup code
